Improve clang-tidy runtimes (#21)
* Update .clang-format-include/ignore and add .clang-tidy-include/ignore

* Reorganize include structure in headers. Hopefully this will improving compilation.

* Run clang-format

* Continue refactoring to reduce the need to include Eigen and OpenFST headers compilation units.

* Move Eigen matrix dependncies to their own header.

* Run clang-format

* Suppress a clang-tidy warning remove static keywords.

* Mark dependencies as system so clang-tidy ignores them.

* Fix or silence warnings produce by clang-tidy

* Silence error on github actions

* Fix "liblibcoati" being used as the file name of the library.

* Simpliy how fstlib_dep is marked as a system dependency.

+++ src/include/coati/semiring.hpp
@@ -47,7 +47,9 @@ class linear {
     static constexpr float times(float x, float y, Args... args) {
         return times(times(x, y), args...);
     }
-    static constexpr float power(float x, size_t y) { return std::pow(x, y); }
+    static constexpr float power(float x, size_t y) {
+        return static_cast<float>(std::pow(x, y));
+    }
     static constexpr float zero() { return 0.0f; }
     static constexpr float one() { return 1.0f; }
 };
